+++ nf_core/subworkflows/__init__.py
@@ -1,0 +1,4 @@
+from .create import SubworkflowCreate
+from .install import SubworkflowInstall
+from .subworkflows_test import SubworkflowsTest
+from .test_yml_builder import SubworkflowTestYmlBuilder